refactor: create a types folder for better organization
