+++ packages/mobile/features/configure_workout/ConfigureWorkout.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { View, TextInput, Button, StyleSheet } from 'react-native';
 import i18nString, { Messages } from '@common/i18n';
 import type { CreateWorkoutNavigationProp } from '@mobile/features/configure_workout/CreateWorkoutScreen';
-import type { Exercise } from '@mobile/stores/workoutSlice';
+import type { Exercise } from '@mobile/types/exercises';
 
 const styles = StyleSheet.create({
   row: {
