+++ packages/mobile/features/home/ActiveWorkout.tsx
@@ -1,11 +1,15 @@
 import React from 'react';
 import { Text } from 'react-native';
 import i18nString, { Messages } from '@common/i18n';
-import type { CurrentWorkoutProps } from '@mobile/types/exercises';
+import type { Exercise } from '@mobile/types/exercises';
+
+interface ActiveWorkoutProps {
+  currentWorkout?: Exercise;
+}
 
 const ActiveWorkout = ({
   currentWorkout,
-}: CurrentWorkoutProps): React.JSX.Element => {
+}: ActiveWorkoutProps): React.JSX.Element => {
   return (
     <Text>
       Current Workout: {currentWorkout?.name ?? i18nString(Messages.Idle)}
